feat: Consolidate page components and logic
Move component logic from dedicated View files into their respective Page components. This reduces duplication and simplifies component management. Also, update statistics calculation for better accuracy.
