+++ components/index.ts
@@ -1,8 +1,6 @@
 export { default as ConfirmationModal } from './ConfirmationModal';
 export { default as LinkModal } from './LinkModal';
 export { default as MarkdownInput } from './MarkdownInput';
-export { default as SettingsView } from './SettingsView';
-export { default as StatsView } from './StatsView';
 export { default as SubtaskModal } from './SubtaskModal';
 export { default as TaskItem } from './TaskItem';
 export { default as TodaySubtaskItem } from './TodaySubtaskItem';
